+++ bin/slipcast.js
@@ -3,6 +3,7 @@
 
 const build = require('../scripts/build');
 const compress = require('../scripts/compress');
+const eject = require('../scripts/eject');
 const minimist = require('minimist');
 const watch = require('../scripts/watch');
 
@@ -21,6 +22,8 @@ if (process.argv.length <= 2 || command.build) {
   }
 } else if (command.compress) {
   compress();
+} else if (command.eject) {
+  eject();
 } else if (command.watch) {
   watch();
 }